06_aks_microservices_azure/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
azure-storage-queue>=12.11.0
azure-functions>=1.20.0
azure-identity>=1.16.0
openai>=1.30.0
langchain>=0.2.0
python-dotenv>=1.0.0
httpx>=0.27.0
